HC-BackEnd/requirements.txt
Flask
Pillow
numpy
opencv-python
Torch
pywebview 
flask_cors
scikit-learn
pandas
openpyxl
pyinstaller